+++ r/R/json.R
@@ -87,19 +87,6 @@ json_table_reader.default <- function(file,
   parse_options = json_parse_options(),
   ...
 ){
-  json_table_reader(fs::path_abs(file),
-    read_options = read_options,
-    parse_options = parse_options,
-    ...
-  )
-}
-
-#' @export
-`json_table_reader.fs_path` <- function(file,
-  read_options = json_read_options(),
-  parse_options = json_parse_options(),
-  ...
-){
   json_table_reader(ReadableFile(file),
     read_options = read_options,
     parse_options = parse_options,